Create timecost_pipeline.go
While studying the pipeline pattern of Go, implemented the mock cost of running the pipeline code vs. sequential code. We have the pipeline with three stages:
multiply -> add -> multiply, which is fed by the channel of 4 integers. Each multiply step costs 3ms(time.Sleep) and add 4ms.
The total cost of the pipeline will be 16ms, which is the total cost of the most expensive task here: add.