Create gotosearch.go
This program is about working with maps. Creating slice of strings in order to keep the order of keys in the map consistent. The program uses recursion to fall through the prerequisite dependencies of some CS courses. Once it finds the course that has not prereq it traces back to the next sibling or parent in the prereq recursive tree.  It then prints out all the courses stored in the string slice order.